ENH: Eliminate the differences between MetaVesselTube and MetaTube
These classes differed only by SetArtery(true/false) function once
ridgeness, medialness, and other non-vessel-specific measures were
moved into the MetaTube class.   With this one member function being
the only difference, it was decided that Arterial is a generic enough
term to also be moved - thereby eliminating the uniqness of the
MetaVesselTube class.   Here it is kept as a facade class of MetaTube.
