+++ src/metaTube.cxx
@@ -20,6 +20,7 @@
 #include <cstdio>
 #include <string>
 
+
 #if (METAIO_USE_NAMESPACE)
 namespace METAIO_NAMESPACE {
 #endif
@@ -39,7 +40,19 @@ TubePnt(int dim)
     m_V2[i]= 0;
     m_T[i]= 0;
     }
+  m_Alpha1=0;
+  m_Alpha2=0;
+  m_Alpha3=0;
   m_R=0;
+  m_Medialness=0;
+  m_Ridgeness=0;
+  m_Branchness=0;
+  m_Curvature=0;
+  m_Levelness=0;
+  m_Roundness=0;
+  m_Intensity=0;
+  m_Mark=false;
+
   //Color is red by default
   m_Color[0]=1.0f;
   m_Color[1]=0.0f;
@@ -62,7 +75,10 @@ MetaTube::
 MetaTube()
 :MetaObject()
 {
-  if(META_DEBUG) std::cout << "MetaTube()" << std::endl;
+  if(META_DEBUG)
+    {
+    std::cout << "MetaTube()" << std::endl;
+    }
   Clear();
 }
 
@@ -71,19 +87,25 @@ MetaTube::
 MetaTube(const char *_headerName)
 :MetaObject()
 {
-  if(META_DEBUG)  std::cout << "MetaTube()" << std::endl;
+  if(META_DEBUG)
+    {
+    std::cout << "MetaTube()" << std::endl;
+    }
   Clear();
   Read(_headerName);
 }
 
 
 MetaTube::
-MetaTube(const MetaTube *_tube)
+MetaTube(const MetaTube *_Tube)
 :MetaObject()
 {
-  if(META_DEBUG)  std::cout << "MetaTube()" << std::endl;
+  if(META_DEBUG)
+    {
+    std::cout << "MetaTube()" << std::endl;
+    }
   Clear();
-  CopyInfo(_tube);
+  CopyInfo(_Tube);
 }
 
 
@@ -91,7 +113,10 @@ MetaTube::
 MetaTube(unsigned int dim)
 :MetaObject(dim)
 {
-  if(META_DEBUG) std::cout << "MetaTube()" << std::endl;
+  if(META_DEBUG)
+    {
+    std::cout << "MetaTube()" << std::endl;
+    }
   Clear();
 }
 
@@ -99,14 +124,14 @@ MetaTube(unsigned int dim)
 MetaTube::
 ~MetaTube()
 {
-  // Delete the list of pointers to tubes.
+  // Delete the list of pointers to Tubes.
   PointListType::iterator it = m_PointList.begin();
   while(it != m_PointList.end())
-{
+    {
     TubePnt* pnt = *it;
     ++it;
     delete pnt;
-}
+    }
   m_PointList.clear();
   M_Destroy();
 }
@@ -116,15 +141,17 @@ void MetaTube::
 PrintInfo() const
 {
   MetaObject::PrintInfo();
-  std::cout << "ParentPoint = " << m_ParentPoint << std::endl;
+  std::cout << "ParentPoint = " << m_ParentPoint
+                      << std::endl;
   if(m_Root)
     {
     std::cout << "Root = " << "True" << std::endl;
     }
   else
     {
-    std::cout << "Root = " << "True" << std::endl;
+    std::cout << "Root = " << "False" << std::endl;
     }
+  std::cout << "Artery = " << m_Artery << std::endl;
   std::cout << "PointDim = " << m_PointDim << std::endl;
   std::cout << "NPoints = " << m_NPoints << std::endl;
   char str[255];
@@ -177,6 +204,19 @@ Root() const
 }
 
 
+void MetaTube::
+Artery(bool artery)
+{
+  m_Artery = artery;
+}
+
+bool MetaTube::
+Artery() const
+{
+  return m_Artery;
+}
+
+
 void  MetaTube::
 ParentPoint(int parentpoint)
 {
@@ -189,34 +229,39 @@ ParentPoint() const
   return m_ParentPoint;
 }
 
-/** Clear tube information */
+/** Clear Tube information */
 void MetaTube::
 Clear()
 {
-  if(META_DEBUG) std::cout << "MetaTube: Clear" << std::endl;
+  if(META_DEBUG)
+    {
+    std::cout << "MetaTube: Clear" << std::endl;
+    }
 
   MetaObject::Clear();
 
   strcpy(m_ObjectTypeName,"Tube");
+  strcpy(m_ObjectSubTypeName,"");
 
-  // Delete the list of pointers to tubes.
+  // Delete the list of pointers to Tubes.
   PointListType::iterator it = m_PointList.begin();
   while(it != m_PointList.end())
-{
+    {
     TubePnt* pnt = *it;
     ++it;
     delete pnt;
-}
+    }
   m_PointList.clear();
 
   m_ParentPoint= -1;
   m_Root = false;
+  m_Artery = true;
   m_NPoints = 0;
-  strcpy(m_PointDim, "x y z r v1x v1y v1z v2x v2y v2z tx ty tz red green blue alpha id");
+  strcpy(m_PointDim, "x y z r rn mn bn cv lv ro in mk v1x v1y v1z v2x v2y v2z tx ty tz a1 a2 a3 red green blue alpha id");
   m_ElementType = MET_FLOAT;
 }
 
-/** Destroy tube information */
+/** Destroy Tube information */
 void MetaTube::
 M_Destroy()
 {
@@ -227,7 +272,11 @@ M_Destroy()
 void MetaTube::
 M_SetupReadFields()
 {
-  if(META_DEBUG) std::cout << "MetaTube: M_SetupReadFields" << std::endl;
+  if(META_DEBUG)
+    {
+    std::cout << "MetaTube: M_SetupReadFields"
+                        << std::endl;
+    }
 
   MetaObject::M_SetupReadFields();
 
@@ -241,6 +290,10 @@ M_SetupReadFields()
 
   mF = new MET_FieldRecordType;
   MET_InitReadField(mF, "Root", MET_STRING, false);
+  m_Fields.push_back(mF);
+
+  mF = new MET_FieldRecordType;
+  MET_InitReadField(mF, "Artery", MET_STRING, false);
   m_Fields.push_back(mF);
 
   mF = new MET_FieldRecordType;
@@ -285,6 +338,19 @@ M_SetupWriteFields()
     m_Fields.push_back(mF);
     }
 
+  if(m_Artery)
+    {
+    mF = new MET_FieldRecordType;
+    MET_InitWriteField(mF, "Artery", MET_STRING, strlen("True"), "True");
+    m_Fields.push_back(mF);
+    }
+  else
+    {
+    mF = new MET_FieldRecordType;
+    MET_InitWriteField(mF, "Artery", MET_STRING, strlen("False"), "False");
+    m_Fields.push_back(mF);
+    }
+
   if(strlen(m_PointDim)>0)
     {
     mF = new MET_FieldRecordType;
@@ -311,18 +377,21 @@ M_Read()
 {
   if(META_DEBUG)
     {
-    std::cout << "MetaTube: M_Read: Loading Header" << std::endl;
+    std::cout << "MetaTube: M_Read: Loading Header"
+                        << std::endl;
     }
 
   if(!MetaObject::M_Read())
     {
-    std::cout << "MetaTube: M_Read: Error parsing file" << std::endl;
+    std::cout << "MetaTube: M_Read: Error parsing file"
+                        << std::endl;
     return false;
     }
 
   if(META_DEBUG)
     {
-    std::cout << "MetaTube: M_Read: Parsing Header" << std::endl;
+    std::cout << "MetaTube: M_Read: Parsing Header"
+                        << std::endl;
     }
 
   MET_FieldRecordType * mF;
@@ -338,14 +407,28 @@ M_Read()
   if(mF->defined)
     {
     if(*((char *)(mF->value)) == 'T'
-      || *((char*)(mF->value)) == 't'
-      || *((char*)(mF->value)) == '1')
+       || *((char*)(mF->value)) == 't'
+       || *((char*)(mF->value)) == '1')
       {
       m_Root = true;
       }
     else
       {
       m_Root = false;
+      }
+    }
+
+  m_Artery = true;
+  mF = MET_GetFieldRecord("Artery", &m_Fields);
+  if(mF->defined)
+    {
+    if(*((char *)(mF->value)) == 'T' || *((char*)(mF->value)) == 't')
+      {
+      m_Artery = true;
+      }
+    else
+      {
+      m_Artery = false;
       }
     }
 
@@ -369,6 +452,14 @@ M_Read()
     posDim[i] = -1;
     }
   int posR = -1;
+  int posRn = -1;
+  int posMn = -1;
+  int posBn = -1;
+  int posCv = -1;
+  int posLv = -1;
+  int posRo = -1;
+  int posIn = -1;
+  int posMk = -1;
   int posV1x = -1;
   int posV1y = -1;
   int posV1z = -1;
@@ -378,6 +469,9 @@ M_Read()
   int posTx = -1;
   int posTy = -1;
   int posTz = -1;
+  int posA1 = -1;
+  int posA2 = -1;
+  int posA3 = -1;
   int posRed = -1;
   int posGreen = -1;
   int posBlue = -1;
@@ -390,7 +484,8 @@ M_Read()
 
   if(META_DEBUG)
     {
-    std::cout << "MetaTube: Parsing point dim" << std::endl;
+    std::cout << "MetaTube: Parsing point dim"
+                        << std::endl;
     }
 
   int j;
@@ -418,6 +513,39 @@ M_Read()
       !strcmp(pntVal[j], "radius") || !strcmp(pntVal[j], "Radius"))
       {
       posR = j;
+      }
+
+    if(!strcmp(pntVal[j], "rn") || !strcmp(pntVal[j], "RN"))
+      {
+      posRn = j;
+      }
+    if(!strcmp(pntVal[j], "mn") || !strcmp(pntVal[j], "MN"))
+      {
+      posMn = j;
+      }
+    if(!strcmp(pntVal[j], "bn") || !strcmp(pntVal[j], "BN"))
+      {
+      posBn = j;
+      }
+    if(!strcmp(pntVal[j], "cv") || !strcmp(pntVal[j], "CV"))
+      {
+      posCv = j;
+      }
+    if(!strcmp(pntVal[j], "lv") || !strcmp(pntVal[j], "LV"))
+      {
+      posLv = j;
+      }
+    if(!strcmp(pntVal[j], "Ro") || !strcmp(pntVal[j], "RO"))
+      {
+      posRo = j;
+      }
+    if(!strcmp(pntVal[j], "in") || !strcmp(pntVal[j], "IN"))
+      {
+      posIn = j;
+      }
+    if(!strcmp(pntVal[j], "mk") || !strcmp(pntVal[j], "MK"))
+      {
+      posMk = j;
       }
     if(!strcmp(pntVal[j], "v1x"))
       {
@@ -455,6 +583,19 @@ M_Read()
       {
       posTz = j;
       }
+    if(!strcmp(pntVal[j], "a1"))
+      {
+      posA1 = j;
+      }
+    if(!strcmp(pntVal[j], "a2"))
+      {
+      posA2 = j;
+      }
+    if(!strcmp(pntVal[j], "a3"))
+      {
+      posA3 = j;
+      }
+
     if(!strcmp(pntVal[j], "red"))
       {
       posRed = j;
@@ -484,7 +625,7 @@ M_Read()
     }
   delete [] pntVal;
 
-  float v[30];
+  float v[50];
 
   if(m_Event)
     {
@@ -507,8 +648,8 @@ M_Read()
                 << std::endl;
       std::cout << "   ideal = " << readSize
                 << " : actual = " << gc << std::endl;
-      delete [] posDim;
       delete [] _data;
+      delete [] posDim;
       return false;
       }
 
@@ -529,7 +670,19 @@ M_Read()
           }
         MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
         i+=sizeof(float);
-        pnt->m_X[d] = td;
+        pnt->m_X[d] = (float)td;
+        }
+
+        {
+        float td;
+        char * num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_R = (float)td;
         }
 
         {
@@ -541,7 +694,96 @@ M_Read()
           }
         MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
         i+=sizeof(float);
-        pnt->m_R = td;
+        pnt->m_Ridgeness = td;
+        }
+
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Medialness = (float)td;
+        }
+
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Branchness = (float)td;
+        }
+
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Curvature = (float)td;
+        }
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Levelness = (float)td;
+        }
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Roundness = (float)td;
+        }
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Intensity = (float)td;
+        }
+
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+
+        if((float)td == 1.0)
+          {
+          pnt->m_Mark = true;
+          }
+        else
+          {
+          pnt->m_Mark = false;
+          }
         }
 
       for(d = 0; d < m_NDims; d++)
@@ -586,6 +828,43 @@ M_Read()
         pnt->m_T[d] = (float)td;
         }
 
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Alpha1 = (float)td;
+        }
+
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Alpha2 = (float)td;
+        }
+
+      if(m_NDims>=3)
+        {
+        float td;
+        char * const num = (char *)(&td);
+        for(k=0;k<sizeof(float);k++)
+          {
+          num[k] = _data[i+k];
+          }
+        MET_SwapByteIfSystemMSB(&td,MET_FLOAT);
+        i+=sizeof(float);
+        pnt->m_Alpha3 = (float)td;
+        }
+
       for(d=0; d<4; d++)
         {
         float td;
@@ -606,9 +885,9 @@ M_Read()
           {
           num[k] = _data[i+k];
           }
-        MET_SwapByteIfSystemMSB(&id,MET_INT);
+        MET_SwapByteIfSystemMSB(&id,MET_FLOAT);
         i+=sizeof(int);
-        pnt->m_ID = id;
+        pnt->m_ID=id;
         }
 
       m_PointList.push_back(pnt);
@@ -638,6 +917,46 @@ M_Read()
         }
 
       pnt->m_R = v[posR];
+
+      if(posMn >= (int)0 && posMn < pntDim)
+        {
+        pnt->m_Medialness = v[posMn];
+        }
+
+      if(posRn >= (int)0 && posRn < pntDim)
+        {
+        pnt->m_Ridgeness = v[posRn];
+        }
+
+      if(posBn >= (int)0 && posBn < pntDim)
+        {
+        pnt->m_Branchness = v[posBn];
+        }
+
+      if(posCv >= (int)0 && posCv < pntDim)
+        {
+        pnt->m_Curvature = v[posCv];
+        }
+
+      if(posLv >= (int)0 && posLv < pntDim)
+        {
+        pnt->m_Levelness = v[posLv];
+        }
+
+      if(posRo >= (int)0 && posRo < pntDim)
+        {
+        pnt->m_Roundness = v[posRo];
+        }
+
+      if(posIn >= (int)0 && posIn < pntDim)
+        {
+        pnt->m_Intensity = v[posIn];
+        }
+
+      if(posMk >= 0 && posMk < pntDim)
+        {
+        pnt->m_Mark = (v[posMk] > 0) ? true:false;
+        }
 
       if(posV1x>=0 && posV1x<pntDim)
         {
@@ -675,6 +994,18 @@ M_Read()
           pnt->m_T[2] = v[posTz];
           }
         }
+      if(posA1 >= 0 && posA1<pntDim)
+        {
+        pnt->m_Alpha1 = v[posA1];
+        }
+      if(posA2 >= 0 && posA2<pntDim)
+        {
+        pnt->m_Alpha2 = v[posA2];
+        }
+      if(posA3 >= 0 && posA3<pntDim)
+        {
+        pnt->m_Alpha3 = v[posA3];
+        }
 
       if(posRed >= 0 && posRed < pntDim)
         {
@@ -705,10 +1036,14 @@ M_Read()
       }
 
 
-    char c = ' ';
-    while( (c!='\n') && (!m_ReadStream->eof()))
+    const std::string objectType = MET_ReadType(*m_ReadStream);
+    if(objectType.empty())
       {
-      c = static_cast<char>(m_ReadStream->get());// to avoid unrecognize charactere
+      char c = ' ';
+      while( (c!='\n') && (m_ReadStream->good()))
+        {
+        c = static_cast<char>(m_ReadStream->get());// to avoid unrecognized characters
+        }
       }
     }
 
@@ -717,7 +1052,7 @@ M_Read()
     m_Event->StopReading();
     }
 
-  delete [] posDim;
+  delete []posDim;
   return true;
 }
 
@@ -739,11 +1074,12 @@ M_Write()
 
   if(!MetaObject::M_Write())
     {
-    std::cout << "MetaTube: M_Read: Error parsing file" << std::endl;
+    std::cout << "MetaTube: M_Read: Error parsing file"
+                        << std::endl;
     return false;
     }
 
-  /** Then copy all tubes points */
+  /** Then copy all Tubes points */
   if(m_BinaryData)
     {
     PointListType::const_iterator it = m_PointList.begin();
@@ -751,65 +1087,101 @@ M_Write()
     int elementSize;
     MET_SizeOfType(m_ElementType, &elementSize);
 
-    const unsigned int writeSize = m_NPoints*(m_NDims*(2+m_NDims)+10)*elementSize;
-
-    char* data = new char[writeSize];
-
-    memset( data, '\0', writeSize );
-
+    char* data = new char[(m_NDims*(2+m_NDims)+10)*m_NPoints*elementSize];
     int i=0;
     int d;
     while(it != itEnd)
       {
       for(d = 0; d < m_NDims; d++)
         {
-        float pntX = (*it)->m_X[d];
-        MET_SwapByteIfSystemMSB(&pntX,MET_FLOAT);
-        MET_DoubleToValue((double)pntX,m_ElementType,data,i++);
+        float x = (*it)->m_X[d];
+        MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+        MET_DoubleToValue((double)x,m_ElementType,data,i++);
         }
 
-      float r = (*it)->m_R;
-      MET_SwapByteIfSystemMSB(&r,MET_FLOAT);
-      MET_DoubleToValue((double)r,m_ElementType,data,i++);
+      float x = (*it)->m_R;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Ridgeness;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Medialness;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Branchness;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Curvature;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Levelness;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Roundness;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Intensity;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Mark;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
 
       for(d = 0; d < m_NDims; d++)
         {
-        float v = (*it)->m_V1[d];
-        MET_SwapByteIfSystemMSB(&v,MET_FLOAT);
-        MET_DoubleToValue((double)v,m_ElementType,data,i++);
+        x = (*it)->m_V1[d];
+        MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+        MET_DoubleToValue((double)x,m_ElementType,data,i++);
         }
 
       if(m_NDims==3)
         {
         for(d = 0; d < m_NDims; d++)
           {
-          float v2 = (*it)->m_V2[d];
-          MET_SwapByteIfSystemMSB(&v2,MET_FLOAT);
-          MET_DoubleToValue((double)v2,m_ElementType,data,i++);
+          x = (*it)->m_V2[d];
+          MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+          MET_DoubleToValue((double)x,m_ElementType,data,i++);
           }
         }
 
       for(d = 0; d < m_NDims; d++)
         {
-        float t = (*it)->m_T[d];
-        MET_SwapByteIfSystemMSB(&t,MET_FLOAT);
-        MET_DoubleToValue((double)t,m_ElementType,data,i++);
+        x = (*it)->m_T[d];
+        MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+        MET_DoubleToValue((double)x,m_ElementType,data,i++);
+        }
+
+      x = (*it)->m_Alpha1;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+      x = (*it)->m_Alpha2;
+      MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+      MET_DoubleToValue((double)x,m_ElementType,data,i++);
+
+      if(m_NDims>=3)
+        {
+        x = (*it)->m_Alpha3;
+        MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+        MET_DoubleToValue((double)x,m_ElementType,data,i++);
         }
 
       for(d=0; d<4; d++)
         {
-        float c = (*it)->m_Color[d];
-        MET_SwapByteIfSystemMSB(&c,MET_FLOAT);
-        MET_DoubleToValue((double)c,m_ElementType,data,i++);
+        x = (*it)->m_Color[d];
+        MET_SwapByteIfSystemMSB(&x,MET_FLOAT);
+        MET_DoubleToValue((double)x,m_ElementType,data,i++);
         }
 
       int id = (*it)->m_ID;
       MET_SwapByteIfSystemMSB(&id,MET_INT);
       MET_DoubleToValue((double)id,m_ElementType,data,i++);
+
       ++it;
       }
 
-    m_WriteStream->write((char *)data,writeSize);
+    m_WriteStream->write((char *)data,
+                         (m_NDims*(2+m_NDims)+10)*m_NPoints*elementSize);
     m_WriteStream->write("\n",1);
     delete [] data;
     }
@@ -827,6 +1199,21 @@ M_Write()
         }
 
       *m_WriteStream << (*it)->m_R << " ";
+      *m_WriteStream << (*it)->m_Ridgeness << " ";
+      *m_WriteStream << (*it)->m_Medialness << " ";
+      *m_WriteStream << (*it)->m_Branchness << " ";
+      *m_WriteStream << (*it)->m_Curvature << " ";
+      *m_WriteStream << (*it)->m_Levelness << " ";
+      *m_WriteStream << (*it)->m_Roundness << " ";
+      *m_WriteStream << (*it)->m_Intensity << " ";
+      if((*it)->m_Mark)
+        {
+        *m_WriteStream << 1 << " ";
+        }
+      else
+        {
+        *m_WriteStream << 0 << " ";
+        }
 
       for(d = 0; d < m_NDims; d++)
         {
@@ -846,6 +1233,14 @@ M_Write()
         *m_WriteStream << (*it)->m_T[d] << " ";
         }
 
+      *m_WriteStream << (*it)->m_Alpha1 << " ";
+      *m_WriteStream << (*it)->m_Alpha2 << " ";
+
+      if(m_NDims>=3)
+        {
+        *m_WriteStream << (*it)->m_Alpha3 << " ";
+        }
+
       for(d=0;d<4;d++)
         {
         *m_WriteStream << (*it)->m_Color[d] << " ";
@@ -858,6 +1253,7 @@ M_Write()
       }
     }
   return true;
+
 }
 
 #if (METAIO_USE_NAMESPACE)
