+++ src/metaVesselTube.h
@@ -16,6 +16,7 @@
 
 #include "metaUtils.h"
 #include "metaObject.h"
+#include "metaTube.h"
 
 #ifdef _MSC_VER
 #pragma warning ( disable: 4251 )
@@ -39,35 +40,18 @@ namespace METAIO_NAMESPACE {
 #endif
 
 class METAIO_EXPORT VesselTubePnt
+  : public TubePnt
 {
 public:
-
   VesselTubePnt(int dim);
   ~VesselTubePnt();
-
-  unsigned int m_Dim;
-  float* m_V1;
-  float* m_V2;
-  float* m_X;
-  float* m_T;
-  float m_Alpha1;
-  float m_Alpha2;
-  float m_Alpha3;
-  float m_R;
-  float m_Medialness;
-  float m_Ridgeness;
-  float m_Branchness;
-  bool  m_Mark;
-  float m_Color[4];
-  int   m_ID;
 };
 
 
 
 
-class METAIO_EXPORT MetaVesselTube : public MetaObject
+class METAIO_EXPORT MetaVesselTube : public MetaTube
 {
-
   /////
   //
   // PUBLIC
@@ -75,12 +59,6 @@ class METAIO_EXPORT MetaVesselTube : public MetaObject
   ////
   public:
 
-   typedef std::list<VesselTubePnt*> PointListType;
-    ////
-    //
-    // Constructors & Destructor
-    //
-    ////
     MetaVesselTube(void);
 
     MetaVesselTube(const char *_headerName);
@@ -92,79 +70,6 @@ class METAIO_EXPORT MetaVesselTube : public MetaObject
     ~MetaVesselTube(void) override;
 
     void PrintInfo(void) const override;
-
-    void CopyInfo(const MetaObject * _object) override;
-
-    //    NPoints(...)
-    //       Required Field
-    //       Number of points which compose the VesselTube
-    void  NPoints(int npnt);
-    int   NPoints(void) const;
-
-    //    PointDim(...)
-    //       Required Field
-    //       Definition of points
-    void        PointDim(const char* pointDim);
-    const char* PointDim(void) const;
-
-    //    Root(...)
-    //       Optional Field
-    //       Set if this VesselTube is a root
-    void  Root(bool root);
-    bool  Root(void) const;
-
-    //    Artery(...)
-    //       Optional Field
-    //       Set if this VesselTube is a root
-    void  Artery(bool artery);
-    bool  Artery(void) const;
-
-
-    //    ParentPoint(...)
-    //       Optional Field
-    //       Set the point number of the parent VesselTube where the branch occurs
-    void  ParentPoint(int parentpoint);
-    int   ParentPoint(void) const;
-
-    void  Clear(void) override;
-
-    PointListType &  GetPoints(void) {return m_PointList;}
-    const PointListType &  GetPoints(void) const {return m_PointList;}
-
-    MET_ValueEnumType ElementType(void) const;
-    void  ElementType(MET_ValueEnumType _elementType);
-
-  ////
-  //
-  // PROTECTED
-  //
-  ////
-  protected:
-
-    bool  m_ElementByteOrderMSB;
-
-    void  M_Destroy(void) override;
-
-    void  M_SetupReadFields(void) override;
-
-    void  M_SetupWriteFields(void) override;
-
-    bool  M_Read(void) override;
-
-    bool  M_Write(void) override;
-
-    int   m_ParentPoint;  // "ParentPoint = "     -1
-
-    bool  m_Root;         // "Root = "            false
-
-    bool  m_Artery;         // "Artery = "            true
-
-    int   m_NPoints;      // "NPoints = "         0
-
-    char m_PointDim[255]; // "PointDim = "       "x y z r"
-
-    PointListType m_PointList;
-    MET_ValueEnumType m_ElementType;
 };
 
 #if (METAIO_USE_NAMESPACE)
